+++ tests/test_checks.py
@@ -123,7 +123,17 @@ class ChecksTestCase(SimpleTestCase):
                     hint="Set DEBUG_TOOLBAR_PANELS to a non-empty list in your "
                     "settings.py.",
                     id="debug_toolbar.W005",
-                )
+                ),
+                Error(
+                    "The Django Debug Toolbar can't be used with tests",
+                    hint="Django changes the DEBUG setting to False when running "
+                    "tests. By default the Django Debug Toolbar is installed because "
+                    "DEBUG is set to True. For most cases, you need to avoid installing "
+                    "the toolbar when running tests. If you feel this check is in error, "
+                    "you can set `DEBUG_TOOLBAR_CONFIG['IS_RUNNING_TESTS'] = False` to "
+                    "bypass this check.",
+                    id="debug_toolbar.W008",
+                ),
             ],
         )
 
@@ -262,28 +272,39 @@ class ChecksTestCase(SimpleTestCase):
             ],
         )
 
-    def test_check_w008_valid(self):
-        settings.DEBUG = True
-        dt_settings.get_config()["IS_RUNNING_TESTS"] = "test"
-        errors = debug_toolbar_installed_when_running_tests_check(None)
-        self.assertEqual(
-            run_checks(),
-            errors,
-            [
-                Error(
-                    "The Django Debug Toolbar can't be used with tests",
-                    hint="Django changes the DEBUG setting to False when running "
-                    "tests. By default the Django Debug Toolbar is installed because "
-                    "DEBUG is set to True. For most cases, you need to avoid installing "
-                    "the toolbar when running tests. If you feel this check is in error, "
-                    "you can set `DEBUG_TOOLBAR_CONFIG['IS_RUNNING_TESTS'] = False` to "
-                    "bypass this check.",
-                    id="debug_toolbar.W008",
-                )
-            ],
-        )
+    # def test_debug_toolbar_installed_when_running_tests(self, mocked_get_config):
+    #     with self.settings(DEBUG=False):
+    #         dt_settings.get_config()["IS_RUNNING_TESTS"] = False
+    #         errors = debug_toolbar_installed_when_running_tests_check(None)
+    #         self.assertEqual(len(errors), 0)
+    #     with self.settings(DEBUG=True):
+    #         dt_settings.get_config()["IS_RUNNING_TESTS"] = True
+    #         errors = debug_toolbar_installed_when_running_tests_check(None)
+    #         self.assertEqual(len(errors), 0)
+    #     with self.settings(DEBUG=True):
+    #         dt_settings.get_config()["IS_RUNNING_TESTS"] = False
+    #         errors = debug_toolbar_installed_when_running_tests_check(None)
+    #         self.assertEqual(len(errors), 0)
+    #     with self.settings(DEBUG=False):
+    #         dt_settings.get_config()["IS_RUNNING_TESTS"] = True
+    #         errors = debug_toolbar_installed_when_running_tests_check(None)
+    #         self.assertEqual(
+    #             errors,
+    #             [
+    #                 Error(
+    #                     "The Django Debug Toolbar can't be used with tests",
+    #                     hint="Django changes the DEBUG setting to False when running "
+    #                     "tests. By default the Django Debug Toolbar is installed because "
+    #                     "DEBUG is set to True. For most cases, you need to avoid installing "
+    #                     "the toolbar when running tests. If you feel this check is in error, "
+    #                     "you can set `DEBUG_TOOLBAR_CONFIG['IS_RUNNING_TESTS'] = False` to "
+    #                     "bypass this check.",
+    #                     id="debug_toolbar.W008",
+    #                 )
+    #             ],
 
-    @patch("debug_toolbar.apps.debug_toolbar_installed_when_running_tests_check", [])
+    # )
+
     def test_debug_toolbar_installed_when_running_tests(self):
         with self.settings(DEBUG=False):
             dt_settings.get_config()["IS_RUNNING_TESTS"] = False